min/최종정리중 ~~~ 251022 17:51

+++ server/scm/src/main/java/com/yedam/scm/vo/MatLotVO.java
@@ -11,8 +11,8 @@ public class MatLotVO {
     private String matId;
     private String lotNo;
     private Date regDate;
-    private Integer initQty;
-    private Integer currQty;
+    private Double initQty;
+    private Double currQty;
     private Date expDate;
     private String warehouse;
     private String empName;
